feat: add toDouble

+++ lib/classes/activity/recordlog/recordlog.dart
@@ -1,22 +1,22 @@
 class RecordLog {
-  int _start = 0;
-  int _end = 0;
+  double _start = 0;
+  double _end = 0;
   late DateTime startDate;
   late DateTime endDate;
   late List<String> notes;
 
   void start() {
     startDate = DateTime.now();
-    _start = DateTime.now().millisecondsSinceEpoch;
+    _start = DateTime.now().millisecondsSinceEpoch.toDouble();
   }
 
   void end() {
     endDate = DateTime.now();
-    _end = DateTime.now().millisecondsSinceEpoch;
+    _end = DateTime.now().millisecondsSinceEpoch.toDouble();
   }
 
   double duration() {
-    double duration = (_end - _start) as double;
+    double duration = _end - _start;
     duration /= 1000;
     return duration;
   }
